rename plugin and plugin function
