+++ android/plugin/src/main/java/plugin/fyberLib/BL_Events.java
@@ -1,4 +1,4 @@
-package plugin.library;
+package plugin.fyberLib;
 
 import java.util.Map;
